workspace mutators auth & error handling

+++ apps/backend/src/workspace/mutators/delete-folder.ts
@@ -1,12 +1,16 @@
 import { RunnerParams } from "../../lib/replicache";
+import { PushAuth } from "../handlers/workspace-push";
 import WorkspaceParty from "../workspace-party";
 import { makeWorkspaceStructureKey } from "@repo/data";
 import { deleteFolderMutation } from "@repo/data/models/workspace/mutators/delete-folder";
 
 export async function deleteFolder(
   party: WorkspaceParty,
-  params: RunnerParams
+  params: RunnerParams,
+  auth: PushAuth
 ) {
+  if (!auth.isOwnerOrAdmin && !auth.isTeamModerator) return;
+
   const args = params.mutation.args as { teamId: string; folderId: string };
   party.workspaceStructure.data = deleteFolderMutation({
     structure: party.workspaceStructure.data,
